requirements.txt
pip~=25.2
attrs~=25.4.0
protobuf~=6.33.0
cryptography~=46.0.3
pillow~=11.3.0
tornado~=6.5.2
Jinja2~=3.1.6
packaging~=25.0
pyparsing~=3.2.5
SQLAlchemy~=2.0.46
streamlit~=1.51.0
pandas~=2.3.3
openpyxl
psycopg2-binary
pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84